tests: split tests option crate in separate test file
